drf_cbv_v1.1/pyproject.toml
[tool.poetry]
name = "drf-cbv"
version = "0.1.0"
description = ""
authors = ["Дамир <barashevdr@student.bmstu.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
djangorestframework = "^3.15.2"
pytest-django = "^4.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
